# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "https://maybets.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://maybets.com/contact-us",
            "email": "tech@maybets.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/incoming/ussd": {
            "post": {
                "description": "This API  receives ussd from Intouch Vas,",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "receives incoming ussd from Intouch Vas",
                "parameters": [
                    {
                        "description": "ussd Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.USSD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "USSD processed successfully",
                        "schema": {
                            "$ref": "#/definitions/models.UssdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing field in the payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Unable to communicate with other micro services",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/templates": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This API sms templates used to send back responses to the customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin - templates"
                ],
                "summary": "Get sms templates",
                "responses": {
                    "200": {
                        "description": "Templates array ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Template"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing field in the payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authorization error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This API updates  sms templates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin - templates"
                ],
                "summary": "Update sms templates",
                "parameters": [
                    {
                        "description": "Template Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Template"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Settings created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing field in the payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authorization error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "required": [
                "error_code",
                "error_message"
            ],
            "properties": {
                "error_code": {
                    "description": "ErrorCode error status code",
                    "type": "integer"
                },
                "error_message": {
                    "description": "ErrorMessage status description",
                    "type": "string"
                }
            }
        },
...
        "models.SuccessResponse": {
            "type": "object",
            "required": [
                "data",
                "status"
            ],
            "properties": {
                "data": {
                    "description": "Message status description",
                    "type": "string"
                },
                "status": {
                    "description": "Status http status code",
                    "type": "integer"
                }
            }
        },
        "models.Template": {
            "type": "object",
            "required": [
                "content",
                "name"
            ],
            "properties": {
                "content": {
                    "description": "Content content of the sms template",
                    "type": "string"
                },
                "name": {
                    "description": "Name name of the sms template",
                    "type": "string",
                    "enum": [
                        "JOIN",
                        "HELP",
                        "BALANCE_QUERY",
                        "SMS_GAMES"
                    ]
                }
            }
        },
        "models.USSD": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "ussd code",
                    "type": "string"
                },
                "input": {
                    "description": "INPUT",
                    "type": "string"
                },
                "level": {
                    "description": "level",
                    "type": "integer"
                },
                "msisdn": {
                    "description": "Msisdn mobile number",
                    "type": "integer"
                },
                "network": {
                    "description": "network",
                    "type": "string"
                },
                "session": {
                    "description": "user Session",
                    "type": "string"
                },
                "text": {
                    "description": "Message sms sent by user",
                    "type": "string"
                }
            }
        },
...
        "models.UssdResponse": {
            "type": "object",
            "properties": {
                "responseType": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api-key",
            "in": "header"
        }
    }
}